Move git utils under utils/git.py
