+++ softlearning/utils/misc.py
@@ -36,26 +36,6 @@ def deep_update(d, *us):
     return d
 
 
-def get_git_rev(path=PROJECT_PATH, search_parent_directories=True):
-    try:
-        import git
-    except ImportError:
-        print(
-            "Warning: gitpython not installed."
-            " Unable to log git rev."
-            " Run `pip install gitpython` if you want git revs to be logged.")
-        return None
-
-    try:
-        repo = git.Repo(
-            path, search_parent_directories=search_parent_directories)
-        git_rev = repo.active_branch.commit.name_rev
-    except TypeError:
-        git_rev = repo.head.object.name_rev
-
-    return git_rev
-
-
 def flatten(unflattened, parent_key='', separator='.'):
     items = []
     for k, v in unflattened.items():
